Fixed format_value and minor tweaks
Changed the format_value function to now return either a single form object or an array of objects (or false if error or empty).
Also updated the readme a bit.

The echoing of shortcode as before resulted in 502 error in my local enviroment and also assumed settings for the field.

Signed-off-by: Jonathan de Jong <jonathan@tigerton.se>